requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.68.1
fastapi-jwt-auth==0.5.0
greenlet==1.1.2
h11==0.12.0
psycopg2-binary==2.9.1
pydantic==1.8.2
PyJWT==1.7.1
python-decouple==3.5
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
Werkzeug==2.0.1
